Add static_cast

+++ examples/all/i2c_example.cc
@@ -22,7 +22,7 @@ static void read_temperature_sensor_value(Mmio<OpenTitanI2c> i2c,
 	i2c->blocking_write(0x48, buf, 1, false);
 	if (i2c->blocking_read(0x48, buf, 2u))
 	{
-		int16_t regValue = (buf[0] << 8) | buf[1];
+		int16_t regValue = static_cast<int16_t>((buf[0] << 8) | buf[1]);
 		int64_t temp     = regValue * 1000000LL / 128;
 		Debug::log("The {} readout is {} microdegrees Celcius", regName, temp);
 	}
